The placeSymbols subprocess now error checks
The placeSymbols subprocess now error checks to make sure the inputted coordinates are usable as coordinates (i.e. they are integers) and also are in range of the dimensions of the grid so will not return an index out of range error which is always nice. The code in that subprocess has been made more efficient (supposedly) through the use of a loop to repeat the code a second time (once for each dimension of the grid).